frontend/api/requirements.txt
fastapi
motor
python-dotenv
email-validator
pydantic[email]
uvicorn
